+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## Unreleased
 ### Fixed
 - Current working directory is not the same as when running programs with Brickman
+### Changed
+- Upload command remembers selected directory for each workspace
 
 ## 0.7.0 - 2017-10-24
 ### Added